dasl::drisl

Module ser

Serialization.

Structs

BoundedCollect

Helper for processing collections.

BufWriter

An in-memory writer.

CollectMap

CBOR RFC-7049 specifies a canonical sort order, where keys are sorted by length first. This was later revised with RFC-8949, but we need to stick to the original order to stay compatible with existing data. We first serialize each map entry (the key and the value) into a buffer and then sort those buffers. Once sorted they are written to the actual output.

CollectSeq

Struct for implementign SerializeSeq.

IoWriter

A writer to work with std::io::Write.

Serializer

A structure for serializing Rust values to DRISL.

Functions

to_vec

Serializes a value to a vector.

to_writer

Serializes a value to a writer.