dasl::drisl::ser

Struct CollectMap

pub struct CollectMap<'a, W> { /* private fields */ }

CBOR RFC-7049 specifies a canonical sort order, where keys are sorted by length first. This was later revised with RFC-8949, but we need to stick to the original order to stay compatible with existing data. We first serialize each map entry (the key and the value) into a buffer and then sort those buffers. Once sorted they are written to the actual output.